pyproject.toml
[project]
name = "article-scheduler"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "feedparser>=6.0.12",
    "python-telegram-bot>=22.6",
]
